move Program default_xx_program to static package
