+++ python/paddle/fluid/tests/unittests/test_arange.py
@@ -16,8 +16,7 @@ from __future__ import print_function
 
 import paddle
 from paddle.fluid import core
-from paddle import Program
-from paddle.static import program_guard
+from paddle.static import program_guard, Program
 import unittest
 import numpy as np
 from op_test import OpTest
