+++ python/paddle/fluid/tests/unittests/test_directory_migration.py
@@ -52,7 +52,9 @@ class TestDirectory(unittest.TestCase):
             'paddle.static.global_scope', 'paddle.static.scope_guard',
             'paddle.static.append_backward', 'paddle.static.gradients',
             'paddle.static.BuildStrategy', 'paddle.static.CompiledProgram',
-            'paddle.static.ExecutionStrategy', 'paddle.static.name_scope',
+            'paddle.static.ExecutionStrategy', 'paddle.static.default_main_program',
+            'paddle.static.default_startup_program', 'paddle.static.Program',
+            'paddle.static.name_scope',
             'paddle.static.program_guard', 'paddle.static.Print',
             'paddle.static.py_func', 'paddle.static.ParallelExecutor',
             'paddle.static.WeightNormParamAttr', 'paddle.static.nn.fc',
@@ -117,7 +119,9 @@ class TestDirectory(unittest.TestCase):
             'paddle.BuildStrategy', 'paddle.CompiledProgram',
             'paddle.ExecutionStrategy', 'paddle.name_scope',
             'paddle.program_guard', 'paddle.Print', 'paddle.py_func',
-            'paddle.ParallelExecutor', 'paddle.WeightNormParamAttr',
+            'paddle.ParallelExecutor', 'paddle.default_main_program',
+            'paddle.default_startup_program', 'paddle.Program',
+            'paddle.WeightNormParamAttr',
             'paddle.declarative.fc', 'paddle.declarative.batch_norm',
             'paddle.declarative.bilinear_tensor_product',
             'paddle.declarative.conv2d', 'paddle.declarative.conv2d_transpose',
