+++ python/paddle/fluid/tests/unittests/test_full_like_op.py
@@ -16,8 +16,7 @@ from __future__ import print_function
 
 import paddle
 import paddle.fluid.core as core
-from paddle import Program
-from paddle.static import program_guard
+from paddle.static import program_guard, Program
 import paddle.compat as cpt
 import unittest
 import numpy as np
