+++ python/paddle/fluid/tests/unittests/test_randperm_op.py
@@ -17,8 +17,7 @@ import numpy as np
 from op_test import OpTest
 import paddle
 import paddle.fluid.core as core
-from paddle import Program
-from paddle.static import program_guard
+from paddle.static import program_guard, Program
 
 
 def check_randperm_out(n, data_np):
